refactor: P3 type cleanup — dedupe VariableType, dethrone functions from dashboard/types
- VariableTableEditor imports VariableType from @/lib/variable-cast instead of
  re-declaring the identical NonNullable<VariableValue["type"]>
- dashboard/types.ts no longer ships runtime functions:
  - formatBytes was unused -> removed
  - formatDuration -> moved to modules/dashboard/utils/format.ts
  - getMethodColor was a diverging Tailwind-class copy -> removed; RunMetadata
    now uses the canonical CSS-var getMethodColor from @/utils (style-based,
    matching DashboardHeader and the design-system "colors via tokens" rule)
- request-builder: replace the `auth as unknown as AuthConfigState` cast in the
  request loader with an explicit per-mode mapper, fixing the latent domain
  `in` -> editor `addTo` rename for apikey auth

type-check clean; 254 tests pass; no new lint errors.
