+++ app/src/modules/dashboard/types.ts
@@ -173,51 +173,3 @@ export interface DisplayMetrics {
 	requests_expected?: number;
 }
 
-// ============================================================================
-// Utility Functions
-// ============================================================================
-
-/**
- * Format duration in human-readable format
- */
-export function formatDuration(ms: number): string {
-	if (ms < 1000) return `${ms}ms`;
-	const seconds = Math.floor(ms / 1000);
-	if (seconds < 60) return `${seconds}s`;
-	const minutes = Math.floor(seconds / 60);
-	const remainingSeconds = seconds % 60;
-	if (minutes < 60) return `${minutes}m ${remainingSeconds}s`;
-	const hours = Math.floor(minutes / 60);
-	const remainingMinutes = minutes % 60;
-	return `${hours}h ${remainingMinutes}m`;
-}
-
-/**
- * Format bytes to human-readable format
- */
-export function formatBytes(bytes: number): string {
-	if (bytes < 1024) return `${bytes} B`;
-	if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-	if (bytes < 1024 * 1024 * 1024) return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-	return `${(bytes / (1024 * 1024 * 1024)).toFixed(1)} GB`;
-}
-
-/**
- * Get HTTP method badge color class
- */
-export function getMethodColor(method: string): string {
-	switch (method.toUpperCase()) {
-		case "GET":
-			return "bg-green-100 text-green-700 dark:bg-green-950 dark:text-green-300";
-		case "POST":
-			return "bg-blue-100 text-blue-700 dark:bg-blue-950 dark:text-blue-300";
-		case "PUT":
-			return "bg-yellow-100 text-yellow-700 dark:bg-yellow-950 dark:text-yellow-300";
-		case "DELETE":
-			return "bg-red-100 text-red-700 dark:bg-red-950 dark:text-red-300";
-		case "PATCH":
-			return "bg-purple-100 text-purple-700 dark:bg-purple-950 dark:text-purple-300";
-		default:
-			return "bg-gray-100 text-gray-700 dark:bg-gray-950 dark:text-gray-300";
-	}
-}
